Dynamic port mapping on ECS
If you use static port mapping, then host port 8080 is always used
when starting containers.  When you deploy a new task definition
(i.e. a new version of the app), ECS fails to start the new
container because the host port is already in use.  Setting the
host port to 0 means ECS allocates the port dynamically.