Adds SBT plugins for building deployable containers
Adds a BuildEnv plugin and an configS3JavaOpts plugin.

The latter retrives configuration in the typesage config library format
from S3 and returns a series of Java option flags suitable for passing
to sbt-native-packager.

The result is that when building a docker image for this application it
will bake configuration from S3 in the Typesafe config format into the
image for the appropriate environment in a format that Finatra
processes as flags.