Provide Jenkins access to volume from host mount
When containerised and accessing jenkins_home via a volume backed on to
a mount on the host, the permissions of the mount persist through to
the container. By default the permissions of an EFS mount as here are
write for root only and consequently jenkins inside the container cannot
write to it's assigned volume.

Adding jenkins as the owner on the host resolves this issue.