Don't use the global ExecutionContext for the transformer
This causes the number of threads in the fork-join pool to be
limited to the number of CPU cores available.  That works fine in
local dev, but in ECS it means one of the threads is never
scheduled.  This new context provides the default parallelism, but
with a larger fork-join pool.