Give the "publish to SNS" Lambda a better name
Because "publish to SNS" only describes what it does, not why it
does it.  Renaming it to "service scheduler" is a better reflection
of its business function.  I tidied up some of the names to be
consistent with our other resources at the same time.

What sort of person thought this was a good name... wait, never mind.