Fix permissions for Lambdas to log to CloudWatch
The new Lambda is unable to log anything to CloudWatch because it
can't create new log streams.  I'm still not sure how the existing
Lambdas were able to create logs without this permission.  Whatever
it was, this allows the new Lambda to log and preserves the ability
of the existing Lambdas to log.