Switch back to building universal/stage
We were seeing issues with builds because we were copying
docker/stage into the Docker image, but rebuilding universal/stage.
The Travis cache was providing a stale binary, hence the never-
changing Docker images.

In 0614ea6, we switched to building docker/stage, but this also
builds a Dockerfile that we don't need.  This patch reverts us back
to using universal/stage, and makes sure that directory is visible
to the Docker build context.