Turn logs off for StartupTests
StartupTests start servers without the overridden modules to connect to the local instances of Dynamo, SQS, etc. Since they try to connect to the real AWS instances but none of the flags with resource names are set (and shouldn't be) they output a lot of errors about not being able to connect to them. The tests still pass as these tests are only testing that all the Guice wiring is correct and the application can startup. So, the error logs are meaningless and make debugging failing tests quite hard. Hence this shitty hack, which does make logs more sane