Fine-tuning works ontology
Adjustments to Works ontology as a result of considering how the Items ontology will look: Shelfmark has been removed from the Works ontology and will appear in Items instead, as hasShelfmark will be a local, Item property rather than one applying to an entire Work.  Also a minor verbal adjustment to the explanation here of what an Item is and how it differs from a Work.