Clarifying "barcode" data property
In order to make clear that this is the Wellcome Library's own barcode for movement control and not e.g. the barcode on a book that encodes its ISBN number, I've renamed this wellcomeBarcode.
ISBN numbers would be another form of identifier to attach to the Work-level record, and ISBN would become another Authority that one could attach to an identifier to indicate its source, basic assumptions, etc.