Use the builtin Terraform behaviour for updating Lambdas
Rather than doing our homebrew tainting of individual Lambdas,
use the hash of the ZIP file and update the Lambda if it changes.
It turns out this is indeed possible with Terraform, but it's
somewhat obliquely documented:

https://github.com/hashicorp/terraform/pull/3825

This should resolve the issues with Lambda triggers being lost,
because the Lambda is only updated, not deleted every time.