Increase the number of uwsgi processes in a single instance
From the uwsgi docs [1]: "There is no magic rule for setting the
number of processes or threads to use".  My current suspicion is
that we're blocking on I/O fetching from S3 and EFS -- so we can
have more processes, which are CPU bound.  This bears out in load
testing, but after 32 ~> 64 processes I didn't see an improvement.

That might be because the current tests don't saturate the process
pool, for now this works but we might want to turn it down again
later.

[1]: http://uwsgi-docs.readthedocs.io/en/latest/ThingsToKnow.html