CALM top of hierarchy example
A large, collection level description coming out of CALM with large amounts of HTML and so forth.  I've done what I can to add backslashes etc in an attempt to make this into valid JSON but there are several points at which I can't work out what the problem is - details are in the markdown file.
There are also a few points at which we need to firm up how we handle a particular field: accession numbers and languages.