Local Make tasks are a bit less chatty
We have some Docker images that rarely changed for make tasks, but
we were rebuilding them _every time_.  That's fine in CI, but
annoying if you're doing it locally -- your console gets spammed
with output from 'docker build'.

Start to modify the Make tasks so they're a bit less chatty.