What if we tried more memory?
We've been seeing a _lot_ of 500s from Loris recently, as well as
spontaneous container reboots -- most of which don't come with an
explanatory log.  Looking at memory graphs for the ECS service,
it's high then drops low around when we have restarts, which seems
like a smoking gun.

Since our t2.xlarge cluster has plenty of spare memory, take
advantage of it.  Memory for everybody!