Adds visibility to Work
In order that visibility states as calculated from sierra data can be
applied.

Items are invisible when deleted is true on Sierra items.
Bibs are invisible when deleted or supressed is true on Sierra bibs.

This PR adds only the status, does not make the change to actually hide
records with visible state false.