catalogue_pipeline: disable the ALB alarms
They've been spamming Slack incessantly since we turned them on
(the alarms, not the services), and since everything stays running we
haven't done anything to fix them -- it seems to be a side-effect of
autoscaling.

Since they don't tell us anything useful and we're not going to "fix"
the services, let's turn off the alarms to reduce noise.