terraform_tracker: Minor refactoring and code cleanup (#1569)
* terraform_tracker: don't hard-code the infra bucket

We're going to rename the bucket soon, so making it a variable will
make it easier to change when we switch.

* terraform_tracker: simplify the code for getting stack name

Since this was written, we've rearranged our state files and given them
a more consistent naming scheme, so this slight fudge can be removed.

* terraform_tracker: improve the whitespace in the Slack message

* terraform_tracker: simplify the 'to_bitly' function

The old version of this code was used in post_to_slack.py, where it
might receive a slash-delimited list of URLs to shorten.  Since it will
only ever get a single URL here, we can remove the list version.