Fix the Sierra bib reader (#1857)
* sierra_adapter: reduce the max capacity of the sierra_reader

We've had issues where early morning windows are being lost (which is
when Sierra does the OAI harvest, and we may have more work to do) --
possibly because of a stampeding herd.  This should reduce the size
of the herd.

* sierra_adapter: remember to UTC-localise our windows

This was causing issues in the build_missing_windows.py script, and
hence replaying windows wasn't working correctly.