Try to fix the flaky Elasticsearch test
The issue was in the test that "updates an already existing index".
We get the occasional 404 "index doesn't exist" error.

The index it uses has the same name as the other tests.  I think
we were hitting a race condition where the index was being created
in this test, then deleted by *another* test before we could use it
a second time.