Storage Stacks!
Allow the entire storage stack to be reproduced easily, whicle keeping persistent resources.

This change should not modify the existing storage service - but will create in parallel all resources except S3 buckets and the hybrid store.

The new stack will be created in the correct VPC but behind api-stage (this mapping can be easily updated when we want to switch).