Adds an S3 bucket for Elastic Cloud snapshots
We currently rely on Elastic Cloud to hold our snapshots for the catalogue API ES cluster.

In order to make working with Catalogue API ES data easy to do locally and to do outside of Elastic Cloud for reduced time and expense, it is useful and more arguably more reliable to hold these in an S3 bucket we have direct access to.

To complete this change we will need to make some changes to the Elastic Cloud configuration.

See: https://www.elastic.co/guide/en/cloud/current/ec-aws-custom-repository.html